pyproject.toml
[tool.poetry]
name = "veloce"
version = "0.1.0"
description = "Rin's Commands API rebuilt using Django"
authors = ["No767 <73260931+No767@users.noreply.github.com>"]
license = "GPL-3.0"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
Django = "^4.1.2"
djangorestframework = "^3.14.0"
Markdown = "^3.4.1"
django-filter = "^22.1"
psycopg2 = "^2.9.5"
PyYAML = "^6.0"
uritemplate = "^4.1.1"
python-dotenv = "^0.21.0"
django-cors-headers = "^3.13.0"
django-prometheus = "^2.2.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^2.20.0"

[tool.isort]
profile = 'black'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
